requirements.txt
streamlit
sqlalchemy
requests
beautifulsoup4
python-dateutil
pandas
numpy